+++ modules/components/GSModal.mjs
@@ -198,11 +198,11 @@ export default class GSModal extends GSElement {
     return me.info(title, message, true, true);
   }
 
-  reset() {
+  reset(data, index = 0) {
     const me = this;
-    me.forms.forEach(f => f.reset());
+    me.forms.forEach(f => {f.reset(); GSDOM.fromObject(f, data);});
     const tab = me.query('GS-TAB');
-    if (tab) tab.index = 0;
+    if (tab && index > -1) tab.index = GSUtil.asNum(index, 0);
   }
 
   /**